relax: split sync.py god module into sync_github / sync_gitlab / sync
All noqa annotations are moved from sync.py — not newly introduced suppressions.
Net change: same set of suppressions, now organized into focused modules.

sync.py (873 lines) had GitHub sync, GitLab sync, ticket upsert, discussion
classification, reviewer sync, and worktree cleanup all mixed together.

Split into three focused modules:
- sync_github.py — GitHub Projects v2 board and reviewer PR discovery
- sync_gitlab.py — GitLab MR upsert, issue labels, merged MR cleanup,
  Slack review permalinks, discussion classification
- sync.py — shared types (SyncResult, MREntry, DiscussionSummary),
  orchestrator (sync_followup), backward-compat re-exports via __all__

Callers and tests importing from teatree.core.sync continue to work
unchanged. Patch targets for cleanup_worktree updated to teatree.core.sync_gitlab.
